feat step3 자동차 경주 게임 - Random 관련한 테스트를 위한 전략패턴 적용 - 자동차 이동에 관련된 메서드 클래스 분리 - 테스트 하기 어려운 Random 관련한 테스트를 위해 FunctionalInterface 사용하여 전략 패턴 적용
todo test 코드 작성
- Random 메서드 전략 패턴을 사용하여 테스트하기 어려운 메소드 stub 데이터로 반환하도록 수정
