+++ src/main/java/racing/core/RacingCarGame.java
@@ -2,18 +2,17 @@ package racing.core;
 
 import racing.domain.RacingCar;
 import racing.domain.RacingRound;
+import racing.rule.MoveRule;
 import racing.vo.RacingCarGamePlayInfo;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 
 public class RacingCarGame {
 
-  public static final int LIMIT_RANDOM_NUMBER = 10;
-  public static final int MORE_THAN_NUMBER_FOR_CAR_MOVE = 4;
   private final RacingCarGamePlayInfo info;
+  private MoveRule moveRule;
 
   private RacingCarGame(RacingCarGamePlayInfo info) {
     this.info = info;
@@ -23,6 +22,11 @@ public class RacingCarGame {
     return new RacingCarGame(info);
   }
 
+  public RacingCarGame setupRule(MoveRule moveRule) {
+    this.moveRule = moveRule;
+    return this;
+  }
+
   public List<RacingRound> play() {
     int totalRound = info.getTotalRound();
     List<RacingCar> racingCars = info.getPlayers();
@@ -30,20 +34,17 @@ public class RacingCarGame {
     List<RacingRound> racingRounds = new ArrayList<>(totalRound);
     for (int i = 0; i < totalRound; i++) {
       for (RacingCar car : racingCars) {
-        if (isMove())
-          car.move();
+        moveOrStop(moveRule, car);
       }
       racingRounds.add(RacingRound.createRecord(racingCars));
     }
     return racingRounds;
   }
 
-  private boolean isMove() {
-    return getRandomNumber() >= MORE_THAN_NUMBER_FOR_CAR_MOVE;
-  }
-
-  private int getRandomNumber() {
-    return new Random().nextInt(LIMIT_RANDOM_NUMBER);
+  private void moveOrStop(MoveRule moveRule, RacingCar car) {
+    if (moveRule.isPossibleMove()) {
+      car.move();
+    }
   }
 
 }
