+++ src/main/java/racing/rule/MoveRule.java
@@ -1,0 +1,7 @@
+package racing.rule;
+
+@FunctionalInterface
+public interface MoveRule {
+
+  boolean isPossibleMove();
+}
